+++ crates/atlaspack_plugin_transformer_js/src/js_transformer/conversion.rs
@@ -685,9 +685,8 @@ fn convert_source_type(source_type: &Option<atlaspack_js_swc_core::SourceType>) 
 
 #[cfg(test)]
 mod tests {
-  use atlaspack_js_swc_core::test_utils::run_swc_core_transform;
-
   use super::*;
+  use atlaspack_js_swc_core::test_utils::run_swc_core_transform;
 
   #[test]
   fn test_is_re_export_all_symbol() {
