+++ crates/atlaspack_plugin_transformer_js/src/js_transformer/conversion/dependency_kind.rs
@@ -40,10 +40,9 @@ pub(crate) fn convert_specifier_type(
 
 #[cfg(test)]
 mod tests {
+  use super::*;
   use atlaspack_js_swc_core::test_utils::run_swc_core_transform;
   use atlaspack_js_swc_core::DependencyKind;
-
-  use super::*;
 
   #[test]
   fn test_import_dependency_kind() {
